Move functions around to avoid circular imports.
